fix type

+++ graphsage/aggregators.py
@@ -13,7 +13,7 @@ class MeanAggregator(Layer):
             name=None, concat=False, **kwargs):
         super(MeanAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -73,7 +73,7 @@ class GCNAggregator(Layer):
             dropout=0., bias=False, act=tf.nn.relu, name=None, concat=False, **kwargs):
         super(GCNAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -123,7 +123,7 @@ class MaxPoolingAggregator(Layer):
             dropout=0., bias=False, act=tf.nn.relu, name=None, concat=False, **kwargs):
         super(MaxPoolingAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -201,7 +201,7 @@ class MeanPoolingAggregator(Layer):
             dropout=0., bias=False, act=tf.nn.relu, name=None, concat=False, **kwargs):
         super(MeanPoolingAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -280,7 +280,7 @@ class TwoMaxLayerPoolingAggregator(Layer):
             dropout=0., bias=False, act=tf.nn.relu, name=None, concat=False, **kwargs):
         super(TwoMaxLayerPoolingAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -367,7 +367,7 @@ class SeqAggregator(Layer):
             dropout=0., bias=False, act=tf.nn.relu, name=None,  concat=False, **kwargs):
         super(SeqAggregator, self).__init__(**kwargs)
 
-        self.dropout = tf.cast(dropout, tf.float64)
+        self.dropout = tf.cast(dropout, tf.float32)
         self.bias = bias
         self.act = act
         self.concat = concat
@@ -410,7 +410,7 @@ class SeqAggregator(Layer):
         initial_state = self.cell.zero_state(batch_size, tf.float32)
         used = tf.sign(tf.reduce_max(tf.abs(neigh_vecs), axis=2))
         length = tf.reduce_sum(used, axis=1)
-        length = tf.maximum(length, tf.constant(1., dtype=tf.float64))
+        length = tf.maximum(length, tf.constant(1., dtype=tf.float32))
         length = tf.cast(length, tf.int32)
 
         with tf.variable_scope(self.name) as scope:
